Move base64_encode and rest_get_random functionst to utils
