+++ src/rest_utils.h
@@ -20,7 +20,14 @@
 #ifndef REST_UTILS_H
 #define REST_UTILS_H
 
+#include <stdint.h>
+#include <stdio.h>
+
+const char *base64_encode(const uint8_t *data, size_t length);
+
 int coap_to_http_status(int status);
+
+size_t rest_get_random(void *buf, size_t buflen);
 
 #endif // REST_UTILS_H
